CSCMETAX-277: [REF] Move REST api code to api/rest/base etc. Move code in tests accordingly. Slightly also rename and move currently existing tests to reflect where the code being tested is located
