+++ src/metax_api/api/rest/base/api_schemas/catalogrecord.json
