+++ src/metax_api/api/rest/base/schemas/contract_schema.json
