+++ src/metax_api/api/rest/base/schemas/datacatalog_schema.json
