+++ src/metax_api/api/rest/base/schemas/file_schema.json
