+++ src/metax_api/api/rest/base/schemas/syke_harvester_dataset_schema.json
