+++ src/metax_api/api/rest/base/serializers/directory_serializer.py
