+++ src/metax_api/api/rest/base/serializers/file_storage_serializer.py
