+++ src/metax_api/api/rest/base/views/directory_view.py
@@ -4,7 +4,7 @@ import logging
 from rest_framework.decorators import detail_route, list_route
 from rest_framework.response import Response
 
-from metax_api.api.base.serializers import DirectorySerializer
+from metax_api.api.rest.base.serializers import DirectorySerializer
 from metax_api.exceptions import Http400, Http501
 from metax_api.models import Directory
 from metax_api.services import FileService
