+++ src/metax_api/api/rest/base/views/schema_view.py
