+++ src/metax_api/api/rest/base/xslt/datacite.xslt
