+++ src/metax_api/tests/utils.py
@@ -11,7 +11,7 @@ test_data_file_path = 'metax_api/tests/testdata/test_data.json'
 
 
 def get_json_schema(model_name):
-    with open(path.dirname(path.realpath(__file__)) + '/../api/base/schemas/%s_schema.json' % model_name) as f:
+    with open(path.dirname(path.realpath(__file__)) + '/../api/rest/base/schemas/%s_schema.json' % model_name) as f:
         return json_load(f)
